refactor(catalog): extract revision comparison
